requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
gradio>=3.50.0
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
scikit-learn>=1.0.0
transformers>=4.30.0

